Move opml loader to itown file
